Moved capsule-gen code to `org.paninij` packages.
me.dwtj.capsules.Capsule -> org.paninij.lang.Capsule
me.dwtj.capsules.CapsuleGenerator -> org.paninij.apt.CapsuleGenerator
