--- org.paninij/capsule-generation/src/main/java/me/dwtj/capsules/Capsule.java
@@ -1,3 +1,0 @@
-package me.dwtj.capsules;
-
-public @interface Capsule { /* no annotation arguments */ }
