+++ org.paninij/capsule-generation/src/main/java/org/paninij/lang/Capsule.java
@@ -1,0 +1,3 @@
+package org.paninij.lang;
+
+public @interface Capsule { /* no annotation arguments */ }
