+++ org.paninij/capsule-generation/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-me.dwtj.capsules.CapsuleGenerator
+org.paninij.apt.CapsuleGenerator
